refactor WindowInPandas code
